requirements.txt
fastapi[all]
uvicorn
httpx
google-genai
python-dotenv
beautifulsoup4
lxml